Extract EditProfileModal as a separate component
